feat(content): created coinbase prep also removed kraken info do to fax only status
