--- public/audio/kraken/open/prepFinal.mp3
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a504e5d5fc1c234d32a29361e044fc53966835dbd960f1225802ec2e46c9e0c6
-size 2378896
