--- public/audio/kraken/open/verifyFinal.mp3
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:625d698abf77f3256c2c673bb32985cd623273c05a213aa40bbb333bb083a8d7
-size 4587852
