+++ src/assets/lessons/tutorials/courseNotes/coinbase/setup/prep.html
@@ -1,0 +1,19 @@
+<div id="content">
+  <h1>Needed Paperwork:</h1>
+  <ol>
+    <li>ID Document</li>
+    <li>Social Security or ITIN</li>
+  </ol>
+  <br>
+  <h1>Other Things Needed:</h1>
+  <ol>
+    <li>Phone Number</li>
+    <li>Email</li>
+    <li>Photo Taking Device</li>
+  </ol>
+  <br>
+  <p>Id Info:</p>
+  <p>https://help.coinbase.com/en/coinbase/getting-started/authentication-and-verification/identity-verification</p>
+  <p>Help Center:</p>
+  <p>https://help.coinbase.com/en/coinbase</p>
+</div>
