--- src/assets/lessons/tutorials/courseNotes/kraken/setup/prep.html
@@ -1,14 +1,0 @@
-<div id="content">
-  <h1>Needed Paperwork:</h1>
-  <ol>
-    <li>ID Document</li>
-    <li>Proof of Residence</li>
-    <li>Social Security or ITIN</li>
-    <li>Face Photo</li>
-  </ol>
-  <br>
-  <p>PaperWork Link:</p>
-  <p>https://support.kraken.com/hc/en-us/articles/360000672203-Document-requirements-for-verification</p>
-  <p>File Type Link:</p>
-  <p>https://support.kraken.com/hc/en-us/articles/234980028</p>
-</div>
